Remove a single file from the index and workspace
Here we begin a simple version of the `rm` command. The most basic use
case we have is to remove files from the index that we no longer want to
be part of the committed tree. This command takes a list of paths, and
removes each of them from the index. And, since Git does this, we also
remove the file from the workspace.

So far this command has no options and does not check for uncommitted
changes -- this behaviour will be added piece by piece in later commits.