Don't remove files with uncommitted changes
The `rm` command tries to be safer than the normal rm(1) program by
refusing to delete data that you can't recover easily. If a file's index
entry differs from its value in HEAD, then the indexed blob is in the
database and will not strictly be lost if the index entry is removed.
However, it will probably not be easy to find the blob in the database
unless we happen to know its ID.

Therefore, `rm` refuses to remove files when they have uncommitted
changes staged in the index. This relies on a new method added to the
`Database` class, which takes a commit/tree ID and a `Pathname`, and
returns the `Database::Entry` for that path in the given tree.