Extract logic for loading a tree listing from the database
The `Repository::Status` class contains logic for loading all the items
in the HEAD tree as a flat map of the form

    { Pathname => Database::Entry }

In order to build the `reset` command, we will need something very
similar to this so that we can fetch the versions of all the files at a
particular commit. Therefore let's extract this logic into the
`Database` class -- `Database#load_list` takes a commit/tree ID and an
optional pathname, and returns the list of files under that path in the
given commit along with their entries.