Launch an editor for preparing commit messages
The workflow for writing commits and merges can be improved by opening a
text editor to let the user input the message. This is particularly
useful for resuming merges, when the user might want to amend the
message after having fixed up the conflicts. This will also let us add
the `--amend` option for `commit`.

The main addition here is the `Editor` class, which takes a pathname,
an environment hash, and a block:

    Editor.edit(path, env) do |editor|
      # ...
    end

This opens the given file for writing, and then yields to the block,
which can write content to the file. `puts` writes content verbatim, and
`note` writes content with a `#` before each line. Once the block exits,
`Editor` launches a text editor using `GIT_EDITOR`, `VISUAL` or `EDITOR`
from the environment. When the text editor program exits, `Editor.edit`
returns the final content of the file, with comments stripped out.

This `Editor` class is used to implement the following use cases. Even
in cases where the user's text editor is not opened, the relevant
content is still written to the edited file.

On `commit`, unless `--message` or `--no-edit` is used, the editor is
opened with some instructions for adding a message to the commit.

On successful `merge`, unless `--message` or `--no-edit` is used, the
editor is opened with instructions for writing a merge message. The
message defaults to `Merge commit '<rev>'`.

On a conflicted `merge`, the given message (`--message` or `--file`) or
the default merge message is written to the `.git/MERGE_MSG` file, with
a list of the conflicted files in the comments. The new method
`Index#conflict_paths` provides this list of files.

When resuming from a merge conflict to write the merge commit, the
message is retrieved from `.git/MERGE_MSG` and some comments reminding
the user they're writing a merge commit are added. The text editor is
then launched and the result becomes the merge commit message.