Add the `--re{use,edit}-message` options to `commit`
A common workflow for making changes to the existing HEAD commit goes:

    $ jit reset --soft @^
    # make changes, amend the index
    $ jit commit --reuse-message ORIG_HEAD

This lets you discard the current HEAD commit, but retrieve its message
so you can use it for the amended commit. Here we add two new options to
the `commit` command:

- `-C`, `--reuse-message` uses the message from the given commit to
  build the new commit. It does not launch the editor.

- `-c`, `--reedit-message` is like `-C`, but opens the editor for
  amending the message before committing.

This just requires a new method in `Command::Commit` for getting the
message from the supplied revision if it was not specified with
`--message` or `--file`.