Set the `committer` field when amending a commit
Until now, all commits we created would have the same `author` and
`committer` information, and so we only passed a single `Author` into
the `Commit` constructor. However, when running `commit --amend`, we
should retain the `author` from the current HEAD, but set `committer` to
the current author, to indicate both who changed it and when. This also
helps ensure that committer timestamps always increase along a history.

This commit adds a new parameter to `Database::Commit` to allow it to
take a distinct committer value, and it amends all sites where commits
are generated to correctly set this field.

We also amend the `Database::Commit#date` method to return the committer
date, not the author one, as this field should always reflect the actual
time the commit was created, and is a better choice for the graph search
heuristics in `RevList`.