Interrupt a cherry-pick on conflicts
Given the changes to `PendingCommit` in the previous commit, we can make
the `cherry-pick` command persist its state if `Merge::Resolve` results
in any conflicts in the index. In this event, we store the picked commit
in `.git/CHERRY_PICK_HEAD` and write the commit's message, plus the list
of conflicted files in comments, to `.git/MERGE_MSG`, before exiting
with an error.

When the command is run with `--continue`, then we load the index and
call a new shared method called `write_cherry_pick_commit`, which
prompts the user to amend the message before building a new commit from
this message and the original commit's author.

The `resume_merge` method dispatches to this functionality if a
cherry-pick is in progress, so the `commit` command will finish the
cherry-pick in the same way that `cherry-pick --continue` does.