Handle conflicts when cherry-picking a range
The `cherry-pick` command can take multiple commits as input, and if any
of them results in a conflict then it will exit and allow the user to
fix the conflicts. The problem is that `PendingCommit` only stores the
currently failing commit in `.git/CHERRY_PICK_HEAD`, so if a commit
part-way through the range fails, the rest of the range is forgotten
about. To properly recover from a conflict, we need to store the rest of
the commits that still need to be picked after the current one.

This task is handled by storing cherry-pick state in `.git/sequencer`,
managed by the `Sequencer` class. This lets a sequence of commits be
written to `.git/sequencer/todo`, it can retrieve the next commit in the
sequence, and drop a commit after it's been successfully picked. The
`cherry-pick` command uses this by giving its initial commit list to
`Sequencer`, and telling this class to dump itself to disk on conflict.

When `--continue` is used, the `Sequencer` can reload its state from
disk and continue the sequence. The code paths mean that starting the
cherry-pick process to begin with, and resuming after a conflict, are
essentially the same -- there's no special error handling code other
than to dump state to disk if the process needs to be interrupted.