Abort and quit cherry-pick sequences
While a cherry-pick sequence is in progress, its state is stored in
`.git/sequencer`. The `cherry-pick` command supports two options for
getting out of the cherry-pick state: `--quit` simply deletes the
pending commit and sequencer state, so the project is left in its
current state but there is no longer a pending commit. `--abort` also
does this, but restores the state of the HEAD, index and workspace back
to the state before the cherry-pick started. This can be used to undo an
entire cherry-pick sequence if a failure is encountered part-way
through, letting you treat it as a transaction.

To support these actions, `Sequencer` stores two new files in
`.git/sequencer`: `abort-safety`, and `head`, both of which initially
contain the commit ID of the current HEAD. `sequencer/head` keeps its
value throughout the process -- this is where we find the original HEAD
value for when `--abort` is used.

`sequencer/abort-safety` is updated after each successfully picked
commit as a safety mechanism. When `--abort` is used, we want to check
the state has not been altered by other commands, i.e. it has not
diverged from the state that the `Sequencer` knows about. If you've made
other commits the `Sequencer` doesn't know about, it shouldn't reset
your HEAD as that would throw those commits away. So, before doing this,
it checks that HEAD matches `abort-safety`, i.e. the current commit is
one created by `cherry-pick`. If so, we perform a hard reset to get the
project back to the state of the commit in `sequencer/head`.