Add a `revert` command to `Sequencer`
The `cherry-pick` command writes commits into `.git/sequencer/todo` as a
line beginning with the word `pick`, followed by the commit ID. The
`Sequencer` needs to be able to store the fact that commits are being
reverted rather than picked, and so we add a new method that stores
commits with the `revert` rather than `pick` instruction.

The `Sequencer#next_command` method will return the instruction as well
as the commit ID, and the caller needs to react appropriately. We'll
shortly implement the `revert` method that reverts a commit instead of
cherry-picking it.