Introduce the `config` command
With the `Config` class we just introduced, we can now implement a
`config` command that reads/writes Git config files. Configuration is
read from `.git/config`, `~/.gitconfig` and `/etc/gitconfig` in order of
precedence; running `config` with just a key returns the last value for
this key in the highest-precedence file. `--get-all` shows all the
recorded values for the key.

Running `config` with a key and value invokes the `set` method to set a
single value. Multiple values can be set using `--add` and replaced
using `--replace-all`. These methods write to `.git/config` by default
but can write to other files depending on the flags that are used.