Select the mainline parent for cherry-pick and revert
Currently, if we cherry-pick or revert a merge commit (one with multiple
parents), then we just use the first parent as the base/target of the
merge. That means the changes brought in from the merged branch will be
picked or reverted.

The `--mainline` option identifies which parent to base the merge on.
Selecting `-m 1` means the first parent will be used as the base, so the
changes coming from the second parent will be picked.

This option needs to be persisted in case of conflicts. We store the
options to `cherry-pick` and `revert` in `.git/sequencer/opts`, which
uses the Git config file format.