Expand the set of colour codes available
To support commands that let their colour output be configured, this
commit adds most of the formatting codes by the name Git uses for them.

This includes a change where, if more than one colour is used in a call
to `fmt`, the second one becomes a background colour by adding 10 to its
SGR code.