Add the `remote` command for adding, removing and listing remote repositories
This commit begins support for remote repositories by introducing the
`remote` command. Adding a remote repository can be done by running:

    $ jit remote add <name> <url>

For example:

    $ jit remote add github ssh://git@github.com/jcoglan/jit.git

This adds the following to the local config file, `.git/config`:

    [remote "github"]
        url = ssh://git@github.com/jcoglan/jit.git
        fetch = +refs/heads/*:refs/remotes/github/*

This instructs us to copy refs from `refs/heads` in the remote
repository to `refs/remotes/github` in our local one.

The `Remotes` class wraps a `Config` object to provide access to store
this state in the `.git/config` file. If `--track` is used, this
overrides the default refspecs put in the `fetch` variable.

The `remove` subcommand deletes a remote by removing its whole section
from the config. Running `remote` with no arguments lists the remotes
along with their URLs for fetching and pushing.