Retrieve the fetch specs and uploader for a remote
In the `fetch` command, we'll need access to a couple of configuration
variables for the remote we're interacting with. The `fetch` variable
tells us which refs to copy from the remote and where to save them in
our local repo. The `uploadpack` variable sets the name of the program
to run on the remote repo to download data from it; this is
`git-upload-pack` by default.

The `fetch_specs` and `uploader` methods added to `Remotes::Remote`
provide access to these values.