Add an `--objects` option to `rev-list` for finding trees and blobs
When we transfer objects from one repository to another, we want to send
all the objects the receiver does not yet have, and *only* those objects
-- to reduce network load, we should try not to send any objects the
receiver already has.

During the `fetch` command, the receiver states which commits it wants
(i.e. those it does not have) as well as those it has, and the sender
should supply all the commits that fill the gap between these sets: the
wanted commits, and their history going back to the commits the receiver
already has. It can find the right commits by constructing a `RevList`
that includes the wanted commits and excludes those the receiver already
has.

But, the sender also needs to supply the *content* of those commits --
the trees and blobs that they contain. In order to avoid sending the
entire project tree, it should avoid sending content objects that are
part of commits the receiver already as. In other words it should only
send trees and blobs that are changed in the commits being sent.

To do this, after the call to `limit_list` that determines the set of
reachable commits, all the trees of commits we marked `uninteresting`
are also marked `uninteresting`. That means that the content of commits
at the edge of the graph, i.e. uninteresting parents of interesting
commits, is excluded from the results, because it's reachable from
commits the receiver has.

After that, the commit queue is traversed as usual, and each commit's
tree is added to the *pending* set. After all commits have been emitted,
this list of pending trees is enumerated, and every object reachable
from these trees is emitted, except those marked as uninteresting.
Deferring this until the end means we can avoid re-traversing trees that
are uninteresting.

This behaviour is enabled by setting the `:objects` option on the
`RevList` class.

Trees and blobs are emitted as `Database::Entry` objects, not `Tree` and
`Blob` objects. All the caller typically needs is the ID and type of
these objects, so we don't need to load the complete blob data.