Reintroduce standard input as an injected dependency
When we began writing tests for this project, we injected dependencies
for the standard input, output and error streams so that we could pass
input to the command under test, and read its output. Once we added
support for setting commit messages using the `--message` option or a
text editor, standard input was not strictly needed any more, so we
removed it.

Now, we're about to introduce new commands that read from stdin: the
`upload-pack` and `receive-pack` commands. For this reason we're
re-injecting stdin as a dependency into `Command` objects.