Load unparsed objects from the database
To prepare a set of objects to send to another repository, we need each
object in its serialised form and we don't need to read its internal
structure. Here we add a method to `Database` called `load_raw` that
reads an object but does not parse it using `Blob`, `Tree` or `Commit`.
Instead it returns a `Raw` object the exposes its type, size and raw
data.

As a side effect, we extract some logic from `read_object` to read the
beginning of the object, and then base `read_object` on that.