Number format for pack object size headers
In the Git pack format [1], each object record begins with a header that
encodes the object's type and its size before zlib compression. The docs
give the format for this header in the somewhat cryptic form:

    n-byte type and length (3-bit type, (n-1)*7+4-bit length)

The `Numbers::VarIntLE` module (variable-length integer, little-endian)
implements the encoding for the size part of the header. The
`(n-1)*7+4-bit` part means the value is encoded as four bits in the
first byte, then seven bits in each subsequent byte. The value is
little-endian, meaning the first byte contains the least-significant
bits.

For example, say we want to encode the number 18,319. Its decimal and
binary representations are:

    decimal:            18319
    binary:   100011110001111

We take the four least-significant (rightmost) bits, `1111`, and because
we have not finished encoding the number, we set the high bit to
indicate more bytes follow.

    10001111

Then we take the next seven least-significant bits, which are `1111000`,
and put them in another byte with the high bit set, so our output is
now:

    10001111  11111000

There are only four bits left in the value now, `1000`, and we put those
in a byte without the high bit set, as it's the last byte of the value.

    10001111  11111000  00001000

In hexadecimal notation, these bytes are:

          8f        f8        08

To decode, we read the bytes back out. 8f is `10001111`, and we take the
low four bits from this:

                        1111

The next byte f8 is `11111000`, and we extract the low seven bits of
this value, and keep going since the high bit is set. We left-shift the
selected bits to place them before the ones we already have:

              1111000   1111

The final byte 08 is `00001000`, and we take the low seven bits of this
value, left-shift them to the front of the others, and stop as the high
bit is not set.

    0001000   1111000   1111

Treating these bits as a binary number recovers the original value
18,319.

The `write` method returns the bytes as an array as the first byte needs
to be further modified before writing the bytes to the file. The type of
the object is stored in the three bits above the four used for the size.
In our example, the bytes were:

    10001111  11111000  00001000

The type of the object, for example 2 (`010`) for a tree, is stored in
the first byte, in the three bits below the high bit.

    10101111  11111000  00001000

[1]: https://git-scm.com/docs/pack-format