Writer for the pack format
When sending objects between repositories, Git uses the pack format [1].
This consists of a header containing the word `PACK`, the version (2),
and the number of objects in the pack. It's followed by all the objects,
each of which is a header containing its type and pre-compression size,
then its data in zlib-compressed form.

The `Pack::Writer` class writes this format to an `IO` stream, which
will be standard output in the case of `upload-pack`. It takes this
stream, a `Database` and some options, and given a `RevList` it will
write the objects in that rev-list to the stream.

The idea is that the caller will construct a `RevList` using the
`:objects` setting to find all the objects the receiver wants but
doesn't have, and then it will pass that list to `Writer` to serialise
it. The `Writer` deals with pulling objects from the database as needed
so they're not all in memory at the same time.

This implementation does not apply any delta compression, so objects
occupy roughly as much space as they would do if stored individually.

[1]: https://git-scm.com/docs/pack-format