Common functions for remote agents
Over the last few commits, we added a few pieces of infrastructure to
support the distribution commands `fetch` and `upload-pack`. The
`remote` command lets us configure remote repositories, the
`Remote::Protocol` class lets us exchange messages with a remote
process, and the changes to `RevList` let us find the objects we want to
send. Finally the `Pack::Writer`, `Pack::Stream` and `Pack::Reader`
classes let us send sets of objects from one repo to another.

This commit begins the building blocks for the control flow of the
`fetch` command. The `RemoteAgent` module contains functions that are
needed by processes that run on the remote repository, which the `fetch`
and `push` commands use to communicate with the remote repo. `fetch`
runs the `upload-pack` command on the remote repo, and `upload-pack`
needs to perform a few tasks that are contained in this module.

First it needs to accept the client's connection and wrap a
`Remotes::Protocol` around it, so that the rest of the functions can use
this `@conn` object to exchange messages.

Second, it needs to detect where the remote Git repository is. It might
be a directory named `.git`, or it might be a directory that contains
Git data but is not named that way. The path of the remote repo is given
as an argument to `upload-pack`, rather than executing the command with
the repo as the working directory. From that argument, we work up the
filesystem until we find a directory that looks like a Git directory, or
contains a `.git` directory that does.

So for example, if we run `upload-pack /path/to/repo`, then we'll try
the following list of directories until we find what we want:

    /path/to/repo
    /path/to/repo/.git
    /path/to
    /path/to/.git
    /path
    /path/.git
    /
    /.git

We stop when we find a directory that contains a file called `HEAD` and
directories called `objects` and `refs`. Enclosing this logic inside the
`repo` method means that all the `Command` code that uses `repo` to
access the current repository can transparently work inside commands
that take the path as an argument, rather than as the current working
directory.

The other task `upload-pack` and other remote agents need to do is send
the current state of their refs to the remote client. For each of their
refs, they send the object ID value and name of the ref. If no refs
exist, then a message is sent with a zero ID and the name
`capabilities^{}`, which is only sent so we have a line to attach the
agent's capability list to.