Send a revision range via the pack protocol
The `SendObjects` module provides some common functionality for sending
a set of objects over the network. Given a list of revisions, which are
object IDs, some prefixed with `^` to exclude them, it constructs a
`RevList` to find all the objects that should be sent, then uses
`Pack::Writer` to serialise them into the pack format.

`@conn` here is an instance of `Remotes::Protocol` that each command --
either `upload-pack` or `push` -- will set up. It means this method can
be used for commands that send their output to different destinations,
using the same API.