Receive a set of objects via the pack protocol
The `ReceiveObjects` module provides the logic needed to process an
incoming pack and store each of its objects to disk. It wraps the input
stream in a `Pack::Stream` and creates a `Pack::Reader` around that. For
each `Pack::Record` the reader yields, we call `Database#store` to save
the object locally.