Check for fast-forward updates
When the `fetch` or `push` command updates a ref, it needs to check if
the proposed update is a fast-forward, i.e. the old ID is for a commit
reachable from the new ID, so that no history is lost. If the `--force`
flag was not used, and the relevant refspec does not have a leading `+`,
then these commands should refuse to perform non-fast-forward updates.

The `FastForward` module provides for this by taking two IDs and
checking if the former is an ancestor of the latter, using
`Merge::CommonAncestors`.