Support shorthand refspecs
In the `fetch` command, we usually use the refspec that's stored in
`.git/config`, which will look something like this:

    +refs/heads/*:refs/remotes/origin/*

This refspec instructs `fetch` to copy each ref from `refs/heads` in the
remote repo to a corresponding file in the local `refs/remotes/origin`
directory, and the leading `+` tells it to force all updates.

For the `push` command, we usually want to push a specific ref, rather
than all of our refs, and it becomes tedious to write

    $ jit push origin refs/heads/master:refs/heads/master

Therefore, refspecs support a number of abbreviations. First, if no
separating `:` is used, then the refspec has the same source and target.
That is, `foo` means the same as `foo:foo`.

Second, the ref names themselves can be abbreviated; if the ref begins
with `heads/` or `remotes/` then it is prefixed with `refs/`, if it
begins with `refs/` then it's left untouched, and all other values are
prefixed with `refs/heads/`. This means the above command can be
shortened to:

    $ jit push origin master

Values are only prefixed with `refs/heads/` if they are valid ref names.
This is because `push` also allows specs like `@~3:master`, meaning it
should write the local value of the revision `@~3` to the remote's
`refs/heads/master`.

Finally, refspecs beginning with `:` indicate there's no source, for
example `:master` means "copy <nil> into `refs/heads/master`". It
effectively means the target ref should be deleted.