Retrieve the `remote.<name>.{push,receivepack}` variables
To support the `push` command, we need to be able to retrieve the
`remote.<name>.push` and `remote.<name>.receivepack` variables from the
configuration. These tell us which refs to push if no arguments are
given, and what program to run on the remote repo instead of
`git-receive-pack`.

Like the other config variables, we can expose these via methods on the
`Remotes::Remote` class.