Display progress of pack transfers
Here we introduce a `Progress` class that can display an updating
read-out of the progress of an operation, counting how many items have
been processed and the percentage completion if an anticipated total is
provided. It can also display how much data has been transferred if
`tick` is called with a numeric argument.

We use this in a few places to make the pack transfer process less
opaque. In `Pack::Writer`, we show the objects being counted and then
written to the pack. In `ReceiveObjects`, we show the objects being
unpacked on the other end of the connection.