Numerical format for delta copy instructions
Objects represented as deltas are composed of two instructions:
`insert`, which is a block of literal data, and `copy` which tells the
decoder to copy a block of data from the base object. These copy
instructions consist of two numbers: the offset and size of the chunk to
copy, occupying up to four and three bytes each.

To encode these numbers, Git effectively turns them into a 56-bit number
by doing `(size << 32) | offset`. This is then encoded in little-endian
form, with any zero bytes removed. A leading byte before this data
indicates which bytes are still present.

The Git documentation [1] gives an example showing the bit pattern of
the leading byte and how it corresponds to the bytes that are present.
To indicate a block represents a copy instruction, Git sets the high bit
on the first byte. Our `PackedInt56LE` module doesn't include this, but
it returns the array of bytes so the caller can add in this extra
information.

[1]: https://git-scm.com/docs/pack-format#_deltified_representation