Serialisation for delta copy and insert operations
The `XDelta` module returns arrays of `Delta::Copy` and `Delta::Insert`
objects. To write these into a pack, these need to be serialised, and
then need to be decoded on the other end.

This commit implements `to_s` and `parse` on these classes to support
this requirement. Copy instructions are serialised using the
`PackedInt56LE` format with the high bit set, and insert instructions
are just a length header (up to 127 or 0x7f) followed by the literal
data.