Load object metadata from the database
When compressing objects in packs, we'd like to get some metadata about
all the objects so they can be sorted, without loading all the objects
into memory. We just need the type and size of each object.

The `Database#load_info` method is similar to `load_raw`, but it only
reads a short amount of the object file in order to parse out the type
and size. It does not return any data from the object.