Compressor for objects in packs
In this commit we introduce, but do not integrate, the
`Pack::Compressor` class. The plan is that `Pack::Writer` will use this
to compress all the objects it wants to send.

`Compressor` accepts all the `Entry` values representing objects the
`Writer` is sending. It sorts the entries by their type, filename and
size, to place likely-similar objects near each other. It then runs a
sliding window over the sorted list to create deltas between pairs of
objects.

Within the sliding window, an object's full data is loaded and an XDelta
index is created for it. For every other object in the window, we try to
create a delta between the pair, and keep the smallest one. The `Delta`
serialisation consists of the sizes of the source and target strings,
encoded using `VarIntLE`, followed by the serialised copy/insert
instructions.

The `Entry` interface exposes the information that `Writer` will need to
write deltified records: the type of the record, the size of the delta,
and the ID of the base object.