Discourage long delta chains that provide little benefit
When compressing a large set of objects, long chains of deltas can form:
structures where a delta record has another delta record as its base,
forming a chain. When a pack is used as primary on-disk storage for
objects, every link in a delta chain adds to the cost of reconstructing
an object, as the base object must be located and expanded first.

So we should only pay that cost if it results in substantial size
savings. Empirically, it seems that objects tend to compress either
extremely well (> 90% size reduction) or very poorly (< 10%). By setting
a minimum size reduction threshold we can avoid creating delta chains
that don't provide much size reduction.

The conditions introduced here reject a delta unless it reduces an
object's size by at least half. If the object has already been assigned
a delta, this threshold is adjusted by the relative depth of the
existing and proposed delta chains, to favour shorter ones.