Extract `Database::Loose` backend
In preparation for using pack files as sources for database objects, we
need to extract the logic from `Database` that's specific to using
individual files to store objects. The `Database::Loose` class contains
all the methods the talk directly to the filesystem, and `Database`
delegates those methods to it.

Extracting this object means that we'll be able to delegate its
interface to other objects with a different implementation, which will
let us introduce support for reading from packs.