Database backend for reading from packs
Having extracted the `Loose` backend from `Database`, we know the
methods we need to implement on classes that could stand in for `Loose`
and perform the same function. Here we implement most of this interface
in `Database::Packed`, which reads from pack files instead of from
individual object files.

To do this, it uses `Pack::Index`, which reads the `.idx` file
associated with a particular `.pack`. This tells us where in the pack
each object ID is, so we can `seek` to that position and ask
`Pack::Reader` to read the record for us. If the record is a delta, then
we recursively read its base object and expand the delta.

`Packed` and `Loose` are not responsible for caching objects or parsing
them into blobs, trees and commits. They just provide a way of accessing
the raw data, and `Database` takes care of the shared logic.