Support multiple database backends
Now that we have `Database::Loose` for reading individual files from
`.git/objects`, and `Database::Packed` for reading from `.pack` files,
we can create a unified object that supports both of these at once.

`Database::Backends` implements the same interface as `Loose` and
`Packed`, but dispatches to all the available data sources. To see if
the database contains an object, we check if any backend has it. To load
an object, we ask each backend in turn until we get a response. To
perform a prefix match, we ask all backends to match the prefix and then
combine their responses.

New objects are still written in loose format -- existing packs are
never modified.