Add `load_info` to the `Database::Packed` interface
When we first implemented the `Database::Packed` class, this method from
the `Loose` interface was not included, as it required new functionality
in the `Pack::Reader` class to read just the header from an object, not
the entire thing.

A delta record's header can give us the size of the original object --
it's contained in the numbers at the start of the delta data. But to get
the object's original type, we need to find the root object of the delta
chain.