Expose and detect remote refs
To properly display remote-tracking refs in the `log` and `branch`
commands, we need to be able to specifically select them, and check
whether any particular ref is a local branch or a remote one.

The `Refs#list_remotes` method returns just the refs in `refs/remotes`.
The `SymRef` class also gains `branch?` and `remote?` methods to tell us
what type it is.