Safely delete branches
When we originally implemented the `--delete` option for `branch`, we
could not implement its full functionality of checking that the branch
is merged with its upstream or HEAD. Therefore we made it do nothing
unless `--force` was also used.

Now, we can implement the check using the `FastForward` module, so that
`branch --delete` succeeds if the branch is an ancestor of its upstream,
or HEAD if it has no upstream.