Show branch divergence in `branch --verbose`
When `branch` is run with the `--verbose` flag, any branch that has an
upstream (a linked remote branch) should display its "divergence" from
its upstream, for example "3 ahead, 2 behind".

The `Divergence` class calculates these numbers by using
`Merge::CommonAncestors` to analyse the history. In verbose mode, we can
then display these numbers to the user. If `-vv` is used, we also
display the name of the upstream branch in blue.