Resolve the `@{upstream}` revision
The revision notation supports expressions like `master@{upstream}`,
which denotes the ID value of the remote-tracking branch that `master`
has as its upstream.

Here we add support for parsing this syntax and resolving the values of
such expressions. The suffix can be abbreviated to `@{u}`, and omitting
the branch name means we should use the current HEAD.